--- Src/Models/Predefines/Coordinates.py
@@ -1,5 +1,0 @@
-import enum
-
-class Coordinates(enum.IntEnum):
-    BARRACKS_X =80
-    BARRACKS_Y =40
