+++ Src/mainLoop.py
@@ -1,5 +1,6 @@
 from pysc2.env import sc2_env
 from pysc2.lib import features
+from Models.BotFile.m_singelton import m_singelton;
 from Models.BotFile.aiBot import aiBot
